+++ app/src/main/java/com/battlelancer/seriesguide/movies/details/MovieDetailsFragment.kt
@@ -76,6 +76,8 @@ import kotlinx.coroutines.withContext
 import org.greenrobot.eventbus.EventBus
 import org.greenrobot.eventbus.Subscribe
 import org.greenrobot.eventbus.ThreadMode
+import org.threeten.bp.Instant
+import org.threeten.bp.ZonedDateTime
 import timber.log.Timber
 
 /**
@@ -119,19 +121,30 @@ class MovieDetailsFragment : Fragment(), MovieActionsContract {
                 isEnabled = false
             }
             // similar movies button
-            buttonMovieSimilar.apply {
-                setOnClickListener {
-                    movieDetails?.tmdbMovie()
-                        ?.title
-                        ?.let {
-                            startActivity(
-                                SimilarMoviesActivity.intent(
-                                    requireContext(),
-                                    tmdbId,
-                                    it
-                                )
+            buttonMovieSimilar.setOnClickListener {
+                movieDetails?.tmdbMovie()
+                    ?.title
+                    ?.let {
+                        startActivity(
+                            SimilarMoviesActivity.intent(
+                                requireContext(),
+                                tmdbId,
+                                it
                             )
-                        }
+                        )
+                    }
+            }
+            buttonMovieCalendar.setOnClickListener {
+                movieDetails?.tmdbMovie()?.also {
+                    val title = it.title
+                    val releaseTimeMs = it.release_date?.time
+                    if (title != null && releaseTimeMs != null) {
+                        ShareUtils.suggestAllDayCalendarEvent(
+                            requireContext(),
+                            title,
+                            releaseTimeMs
+                        )
+                    }
                 }
             }
             buttonMovieCheckIn.setOnClickListener { onButtonCheckInClick() }
@@ -339,10 +352,15 @@ class MovieDetailsFragment : Fragment(), MovieActionsContract {
 
         // release date and runtime: "July 17, 2009 | 95 min"
         val releaseAndRuntime = StringBuilder()
-        tmdbMovie.release_date?.let {
+        val releaseDate = tmdbMovie.release_date
+        releaseDate?.let {
             releaseAndRuntime.append(TimeTools.formatToLocalDate(context, it))
             releaseAndRuntime.append(" | ")
         }
+        // hide create event button if release date is yesterday or older
+        binding.containerMovieButtons.buttonMovieCalendar.isGone =
+            releaseDate == null || Instant.ofEpochMilli(releaseDate.time)
+                .isBefore(ZonedDateTime.now().minusDays(1).toInstant())
         tmdbMovie.runtime?.let {
             releaseAndRuntime.append(getString(R.string.runtime_minutes, it.toString()))
         }